tests/wpt/web-platform-tests/loading/lazyload/original-referrer-policy-applied-tentative.sub.html

<!DOCTYPE html>
<head>
  <title>Deferred iframes and images with loading='lazy' use the original referrer-policy specified at the parse time</title>
  <link rel="author" title="Raj T" href="mailto:rajendrant@chromium.org">
  <link rel="help" href="https://github.com/scott-little/lazyload">
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="common.js"></script>
</head>

<!--
Marked as tentative until https://github.com/whatwg/html/pull/3752 is landed.
-->

<script>
  const below_viewport_iframe = new ElementLoadPromise("below_viewport_iframe");

  // Change the referrer-policy and scroll down to load the deferred elements.
  window.addEventListener("load", () => {
    below_viewport_iframe.element().referrerPolicy = "no-referrer";
    document.getElementById("below_viewport_iframe").scrollIntoView();
  });

  async_test(function(t) {
    below_viewport_iframe.promise.then(
      t.step_func_done(function() {
        // The referer header should be the full URL (as specified in the iframe
        // at parse time), and not the origin (as specified in meta referrer
        // tag) or null (as overridden by iframe referrerpolicy=no-referrer).
        assert_true(below_viewport_iframe.element().contentDocument.body.innerHTML
            .includes("Referer: http://{{host}}:{{ports[http][0]}}/loading/lazyload/"));
      }));
  }, "Test that when deferred iframe is loaded, it uses the referrer-policy specified at parse time.");
</script>

<body>
  <meta name="referrer" content="origin">
  <div style="height:10000px;"></div>
  <iframe id="below_viewport_iframe" src="/xhr/resources/echo-headers.py" loading="lazy" width="200px" height="100px" referrerpolicy="unsafe-url" onload="below_viewport_iframe.resolve();">
  </iframe>
</body>
