third_party/blink/web_tests/external/wpt/loading/lazyload/original-referrer-policy-applied-tentative.sub-expected.txt

main frame - DidStartNavigation
main frame - ReadyToCommitNavigation
main frame - didCommitLoadForFrame
main frame - didReceiveTitle: Deferred iframes and images with loading='lazy' use the original referrer-policy specified at the parse time
frame "<!--framePath //<!--frame0-->-->" - didReceiveTitle: 
main frame - didFinishDocumentLoadForFrame
main frame - didHandleOnloadEventsForFrame
main frame - didFinishLoadForFrame
frame "<!--framePath //<!--frame0-->-->" - BeginNavigation request to 'http://web-platform.test:8001/xhr/resources/echo-headers.py', http method GET
frame "<!--framePath //<!--frame0-->-->" - DidStartNavigation
frame "<!--framePath //<!--frame0-->-->" - ReadyToCommitNavigation
frame "<!--framePath //<!--frame0-->-->" - didCommitLoadForFrame
frame "<!--framePath //<!--frame0-->-->" - didReceiveTitle: 
frame "<!--framePath //<!--frame0-->-->" - didFinishDocumentLoadForFrame
frame "<!--framePath //<!--frame0-->-->" - didHandleOnloadEventsForFrame
frame "<!--framePath //<!--frame0-->-->" - didFinishLoadForFrame
This is a testharness.js-based test.
PASS Test that when deferred iframe is loaded, it uses the referrer-policy specified at parse time.
FAIL Test that when deferred img is loaded, it uses the referrer-policy specified at parse time. assert_unreached: The image load should not fail, by sending the wrong referer header. Reached unreachable code
Harness: the test ran to completion.
